big reorg for better package structure and tests
