+++ data/__init__.py
@@ -5,4 +5,4 @@ _SRC = Path(__file__).resolve().parents[1] / "src"
 if str(_SRC) not in sys.path:
     sys.path.insert(0, str(_SRC))
 
-from speedrunning_plms.models.plm import *  # noqa: F401,F403
+from speedrunning_plms.data import *  # noqa: F401,F403
