+++ data/create_omgprot50_splits.py
@@ -1,33 +1,22 @@
 import argparse
-from datasets import load_dataset, DatasetDict
+import sys
+from pathlib import Path
 
-parser = argparse.ArgumentParser()
-parser.add_argument('--hf_token', type=str, default=None)
+_SRC = Path(__file__).resolve().parents[1] / "src"
+if str(_SRC) not in sys.path:
+    sys.path.insert(0, str(_SRC))
 
-args = parser.parse_args()
+from speedrunning_plms.data.splits import build_omg_prot50_splits, login_if_token, push_splits
 
-if args.hf_token:
-    import huggingface_hub
-    huggingface_hub.login(token=args.hf_token)
 
-data = load_dataset('tattabio/OMG_prot50', split='train').remove_columns('id').shuffle(seed=11)
-#data = data.cast_column('sequence', Value(dtype='string'))
-print(data)
+def main() -> None:
+    parser = argparse.ArgumentParser()
+    parser.add_argument("--hf_token", type=str, default=None)
+    args = parser.parse_args()
+    login_if_token(args.hf_token)
+    data = build_omg_prot50_splits()
+    push_splits(data, "Synthyra/omg_prot50")
 
-data = data.train_test_split(test_size=20000, seed=22)
 
-train = data['train']
-valid = data['test']
-valid = valid.train_test_split(test_size=10000, seed=33)
-test = valid['test']
-valid = valid['train']
-
-data = DatasetDict({
-    'train': train,
-    'valid': valid,
-    'test': test
-})
-
-print(data)
-
-data.push_to_hub('Synthyra/omg_prot50')
+if __name__ == "__main__":
+    main()
