+++ src/speedrunning_plms/flex/__init__.py
@@ -1,0 +1,11 @@
+from speedrunning_plms.flex.mods import (
+    create_score_mod,
+    generate_dilated_sliding_window,
+    visualize_attention_scores,
+)
+
+__all__ = [
+    "create_score_mod",
+    "generate_dilated_sliding_window",
+    "visualize_attention_scores",
+]
