+++ src/speedrunning_plms/flex/mods.py
@@ -1,0 +1,221 @@
+# https://github.com/pytorch-labs/attention-gym/blob/main/attn_gym/mods/softcapping.py
+
+import math
+import numpy as np
+import torch
+from typing import Optional
+from pathlib import Path
+from torch.nn.attention.flex_attention import (
+    _score_mod_signature,
+    _mask_mod_signature,
+    _vmap_for_bhqkv,
+    _ModificationType,
+)
+
+try:
+    from torch._dynamo._trace_wrapped_higher_order_op import TransformGetItemToIndex
+except ImportError:
+    from torch._higher_order_ops.flex_attention import TransformGetItemToIndex
+from contextlib import nullcontext
+
+
+def create_score_mod(
+    query: torch.Tensor,
+    key: torch.Tensor,
+    score_mod: Optional[_score_mod_signature],
+    mask_mod: Optional[_mask_mod_signature],
+    device: str = "cuda",
+    _compile: bool = False,
+    scale: Optional[float] = None,
+    batch_idx: int = 0,
+    head_idx: int = 0,
+) -> torch.Tensor:
+    B = 1
+    H = 1
+    M = query.shape[0]
+    N = key.shape[0]
+
+    b = torch.arange(0, B, device=device) + batch_idx
+    h = torch.arange(0, H, device=device) + head_idx
+    m = torch.arange(0, M, device=device)
+    n = torch.arange(0, N, device=device)
+
+    scale_factor = 1 / math.sqrt(query.size(-1)) if scale is None else scale
+    type = _ModificationType.SCORE_MOD if score_mod is not None else _ModificationType.MASK_MOD
+    if _compile:
+        ctx = nullcontext()
+    else:
+        ctx = TransformGetItemToIndex()
+
+    with ctx:
+        mod_fn = score_mod if type == _ModificationType.SCORE_MOD else mask_mod
+        prefix = (0,) if type == _ModificationType.SCORE_MOD else ()
+        mod = _vmap_for_bhqkv(mod_fn, prefix=prefix)
+        scores = query @ key.transpose(-2, -1)
+        scores *= scale_factor
+        scores = scores.view(1, 1, M, N)
+        if type == _ModificationType.SCORE_MOD:
+            out = mod(scores, b, h, m, n)
+        else:
+            out = mod(b, h, m, n)
+
+    return out
+
+
+def generate_dilated_sliding_window(window_size: int, dilation: int) -> _mask_mod_signature:
+    """Generates a dilated sliding window attention mask.
+    Args:
+        window_size: The size of the sliding window.
+        dilation: The dilation factor for the sliding window.
+
+    Note:
+        Query at position i can only attend to keys within a window of size `window_size`
+        centered around i, where the keys are at positions j such that:
+        * abs(i - j) <= window_size
+        * abs(i - j) % dilation == 0
+    """
+
+    def dilated_sliding_window(b, h, q_idx, kv_idx):
+        diff = torch.abs(q_idx - kv_idx)
+        in_window = diff <= window_size
+        is_dilated = (diff % dilation) == 0
+        return in_window & is_dilated
+
+    dilated_sliding_window.__name__ = f"dilated_sliding_window_{window_size}_dilation_{dilation}"
+    return dilated_sliding_window
+
+
+def _name_to_title(name: str) -> str:
+    title = name.replace("_", " ")
+    title = " ".join(word.capitalize() for word in title.split())
+    return title
+
+
+def visualize_attention_scores(
+    query: torch.Tensor,
+    key: torch.Tensor,
+    score_mod: Optional[_score_mod_signature] = None,
+    mask_mod: Optional[_mask_mod_signature] = None,
+    device: str = "cuda",
+    name: str = "attention_scores",
+    path: Optional[Path] = None,
+    batch_idx: int = 0,
+    head_idx: int = 0,
+    scale: Optional[float] = None,
+):
+    """
+    Generate and save a visualization of attention scores.
+
+    Args:
+        query (Tensor): Query tensor of shape (batch_size, num_heads, seq_len_q, head_dim).
+        key (Tensor): Key tensor of shape (batch_size, num_heads, seq_len_k, head_dim).
+        score_mod (Optional[Callable]): If this is set this will take precedence over the mask_mod.
+        mask_mod (Optional[Callable]): The mask_mod function used to create block_mask
+        device (str): Device to run computations on (default: "cuda").
+        name (str): Base name for the file and title (default: 'attention_scores').
+        path (Path): Path to save the visualization. If None, will be saved to the current working directory.
+        batch_idx (int): Index of the batch to visualize (default: 0).
+        head_idx (int): Index of the head to visualize (default: 0).
+        scale (float): Scale factor to apply to the attention scores. If None, will be set to 1 / sqrt(head_dim).
+
+    Returns:
+        None
+    """
+    import matplotlib.pyplot as plt
+
+    assert score_mod is not None or mask_mod is not None, (
+        "Must provide either score_mod or mask_mod"
+    )
+    query = query[batch_idx, head_idx, :, :]
+    key = key[batch_idx, head_idx, :, :]
+    scores_viz = create_score_mod(
+        query,
+        key,
+        score_mod=score_mod,
+        mask_mod=mask_mod,
+        scale=scale,
+        device=device,
+        batch_idx=batch_idx,
+        head_idx=head_idx,
+    )
+    # If both score_mod and mask_mod are provided, apply both
+    if score_mod is not None and mask_mod is not None:
+        mask_viz = create_score_mod(
+            query,
+            key,
+            score_mod=None,
+            mask_mod=mask_mod,
+            scale=scale,
+            device=device,
+            batch_idx=batch_idx,
+            head_idx=head_idx,
+        )
+        # Apply mask by setting masked positions to -inf
+        scores_viz = torch.where(mask_viz == 0, float("-inf"), scores_viz)
+
+    suffix_title = f"Batch {batch_idx}, Head {head_idx}" if batch_idx != 0 or head_idx != 0 else ""
+
+    fig, ax = plt.subplots(figsize=(12, 10))
+    color = "viridis" if score_mod is not None else "cividis"
+    if score_mod is not None and mask_mod is not None:
+        color = "plasma"
+    im = ax.imshow(scores_viz.cpu().detach()[0, 0, :, :], aspect="auto", cmap=color)
+    fig.colorbar(im)
+
+    title = _name_to_title(name)
+    file_path = Path(name).with_suffix(".png") if path is None else path.with_suffix(".png")
+    ax.set_title(f"{title}\n{suffix_title}", fontsize=20)
+
+    ax.set_xlabel("Key Tokens", fontsize=18)
+    ax.set_ylabel("Query Tokens", fontsize=18)
+
+    # Move y-axis ticks and labels to the top
+    ax.tick_params(axis="x", top=True, labeltop=True, bottom=False, labelbottom=False)
+
+    # Add tick labels if the number of tokens is manageable
+    num_query_tokens, num_kv_tokens = scores_viz.shape[-2:]
+    if num_query_tokens <= 32 and num_kv_tokens <= 32:
+        ax.set_xticks(range(num_kv_tokens))
+        rotation = 45 if num_kv_tokens > 12 else 0
+        ax.set_xticklabels(
+            [f"KV{i}" for i in range(num_kv_tokens)], fontsize=16, rotation=rotation
+        )
+        ax.set_yticks(range(num_query_tokens))
+        ax.set_yticklabels([f"Q{i}" for i in range(num_query_tokens)], fontsize=16)
+        # Align grid with pixel boundaries
+        ax.set_xticks(np.arange(-0.5, num_kv_tokens, 1), minor=True)
+        ax.set_yticks(np.arange(-0.5, num_query_tokens, 1), minor=True)
+        ax.grid(which="minor", color="black", linestyle="-", linewidth=2)
+
+    plt.tight_layout()
+    plt.savefig(file_path, dpi=300, bbox_inches="tight")
+    plt.close(fig)  # Close the figure to free up memory
+
+    print(f"Visualization saved as {file_path}")
+
+
+def main(device: str = "cpu"):
+    """Visualize the attention scores of dilated sliding window mask mod.
+
+    Args:
+        device (str): Device to use for computation.
+    """
+    B, H, SEQ_LEN, HEAD_DIM = 1, 1, 24, 8
+
+    def make_tensor():
+        return torch.ones(B, H, SEQ_LEN, HEAD_DIM, device=device)
+
+    query, key = make_tensor(), make_tensor()
+
+    dilated_sliding_window_mask = generate_dilated_sliding_window(window_size=8, dilation=4)
+    visualize_attention_scores(
+        query,
+        key,
+        mask_mod=dilated_sliding_window_mask,
+        device=device,
+        name="dilated_sliding_window_mask",
+    )
+
+
+if __name__ == "__main__":
+    main()
