+++ src/speedrunning_plms/models/config.py
@@ -1,0 +1,3 @@
+from speedrunning_plms.models.plm import ESMOutput, PLMConfig
+
+__all__ = ["ESMOutput", "PLMConfig"]
