+++ src/speedrunning_plms/models/masks.py
@@ -1,0 +1,3 @@
+from speedrunning_plms.models.plm import precompute_multiresolution_masks
+
+__all__ = ["precompute_multiresolution_masks"]
