+++ src/speedrunning_plms/optim/__init__.py
@@ -1,0 +1,3 @@
+from speedrunning_plms.optim.muon import Muon, zeropower_via_newtonschulz5
+
+__all__ = ["Muon", "zeropower_via_newtonschulz5"]
